chore: Move `react-intl` to shared libs
