+++ apps/flame-defi/app/providers.tsx
@@ -4,7 +4,7 @@ import { wallets as keplrWallets } from "@cosmos-kit/keplr";
 import { wallets as leapWallets } from "@cosmos-kit/leap";
 import { ChainProvider } from "@cosmos-kit/react";
 import { RainbowKitProvider, getDefaultConfig } from "@rainbow-me/rainbowkit";
-import { IntlProvider } from "@repo/ui/intl";
+import { IntlProvider } from "@repo/libs/react-intl";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { assets, chains } from "chain-registry";
 import { WagmiProvider } from "wagmi";
