+++ apps/flame-defi/package.json
@@ -23,6 +23,7 @@
     "@metamask/jazzicon": "^2.0.0",
     "@rainbow-me/rainbowkit": "^2.2.2",
     "@repo/flame-types": "*",
+    "@repo/libs": "*",
     "@repo/ui": "*",
     "@tanstack/react-query": "^5.61.0",
     "@tanstack/react-table": "^8.20.6",
