+++ packages/libs/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@repo/libs",
+  "version": "0.0.0",
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "lint": "eslint . --max-warnings 0",
+    "check-types": "tsc --noEmit"
+  },
+  "exports": {
+    "./*": "./src/*/index.ts"
+  },
+  "devDependencies": {
+    "@repo/eslint-config": "*",
+    "@repo/typescript-config": "*",
+    "typescript": "5.5.4"
+  },
+  "dependencies": {
+    "react-intl": "^7.1.6"
+  }
+}
