+++ packages/ui/package.json
@@ -9,7 +9,6 @@
     "./hooks": "./src/hooks/index.ts",
     "./utils": "./src/utils/index.ts",
     "./icons": "./src/icons/index.ts",
-    "./intl": "./src/intl/index.ts",
     "./types": "./src/types.ts"
   },
   "scripts": {
@@ -45,13 +44,13 @@
     "@radix-ui/react-tabs": "^1.1.3",
     "@radix-ui/react-tooltip": "^1.1.6",
     "@repo/flame-types": "*",
+    "@repo/libs": "*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.471.1",
     "motion": "^12.4.10",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-intl": "^7.1.6",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^1.1.2"
